优化获取ssl证书到期时间接口

+++ app/Models/Ssl.php
@@ -10,9 +10,14 @@ class Ssl extends Base
         $data['pem'] = Cache::get("ssl_pem");
         $data['end_time'] = '';
         if(function_exists('exec') && !empty($data['pem'])){
-            file_put_contents(public_path().'/ssl_pem.txt',$data['pem']);
-            exec("openssl x509 -in ".public_path()."/ssl_pem.txt"." -noout -text",$res);
-            unlink(public_path().'/ssl_pem.txt');
+            $cache_key = 'pem:'.md5($data['pem']);
+            $res = Cache::get($cache_key);
+            if(empty($res)){
+                file_put_contents(public_path().'/ssl_pem.txt',$data['pem']);
+                exec("openssl x509 -in ".public_path()."/ssl_pem.txt"." -noout -text",$res);
+                unlink(public_path().'/ssl_pem.txt');
+                Cache::put($cache_key,$res,60*60*24);
+            }
             foreach ($res as $k =>$v){
                 if(strpos($v,'Not After') !== false){
                     $tmp = strtotime(str_replace('Not After : ','',$v));
